storage_service: force_remove_endpoint of changed ip address
When a host changes its ip address we should force remove
the previous endpoint state since we want only one endpoint
to refer to this host_id.

If the new node that changed its ip address is decommissioned,
the previous node seems as a normal token owner, just in shutdown
status, but it is not longer in the cluster.

Refs scylladb/scylladb#14468
Fixes scylladb/scylladb#13775

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>