token_metadata: remove_endpoint: delete replacing endpoint by value
Currently, we only attempt to remove a replacing endpoint
from _replacing_endpoints by key, i.e. as an existing node,
but it might exist there as a value, i.e. as a replacing node
with consistent-topology-changes.

Regardless, remove_endpoint should leave no trace of the the
removed endpoint in token_metadata and the respective topology
hence the fix is implemented here and not in storage_service.

Fixes #16730

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>