feat: add system.repair_history ttl param
we add an option for repair_history ttl.
I found if I use gcmode='repair' scylla will create many record in table system.repair_history.
And this record will not be deleted, this table will grow without bounds.
And scylla will read this table when scylla launch.
It takes a long time to read system.repair_history will.
This option add ttl to record in system.repair_history when the record be inserted.