utils/logalloc: on non-lsa frees, call decrease_usage() with a negative argument
region_listener::decrease_usage() is supposed to be called with a negative
delta. (I think it's a bad API, but it is what it is). But for non-LSA frees,
it's called with the wrong sign.

I don't know if it has any consequences in practice, but since the bug has
been present for 8 years, it's probably not very important.