mv: handle mismatched base/view replica count caused by RF change
During a ALTER KEYSPACE statement execution where a table with
a view is present, we need to perform tablet migrations for both
tables. These migrations are not synchronized, so at some point
the base may have a different number of replicas than the view.
Because of that, we can't pair them correctly, which can cause
a crash.

This patch adds a workaround for this scenario. If after one
migration we have too many base tablets, we simply do not
generate updates from it, and the new view replica (in case of
RF increase) is pending and will get updates from all other base
replicas.
If after one migration we have too many view replicas, no base
replica will be paired with it, so we add it to the pending
replica list so that it gets an update anyway.
Replication changes only modify the last replica in the list, so
we identify the last entry in the list as the new replica

This patch will also take effect if the base and view replica
counts differ due to some other bug. To track that, a new metric
is added to count such occurrences.

This patch also includes a test for this exact scenario, which
is enforced by an injection.

Fixes https://github.com/scylladb/scylladb/issues/21492