main, view: Pair view builder drain with its start
Before these changes, we only drained the view builder when stopping it,
which was paired with its *construction*. However, since the view
builder depends on group0_service, which is initialized AFTER the view
builder is constructed, that could lead to segmentation faults.
To amend it, we pair draining the view builder with its start, while
leaving the lambda destroying it paired with its construction. This way
we should have a proper initialization order, avoid bugs, and reinforce
our understanding of how Scylla starts up.

Fixes scylladb/scylladb#20772