compaction: set _child_count value for table compaction tasks
Set the _child_count for table tasks of the following compaction
types:
- major;
- cleanup;
- offstrategy;
- upgrade_sstables;
- scrub.
The children of these table tasks span over a single compaction
group. Hence, the _child_count is equal to the number of cgs
in the table.

Reshape and reshard compactions have different structure and will
be covered in the following patches.